pyproject.toml
[project]
name = "python-kursovai-5"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker (>=37.11.0,<38.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_paths = ["."]
addopts = "--reuse-db --cov=. --cov-report=html --cov-report=term-missing"
markers = [
    "slow: slow running tests",
    "unit: unit tests",
    "integration: integration tests"
]
